feat(svc): migrate dmsg-server onto pkg/services framework
Extracts cmd/dmsg/dmsg-server/start's run loop into
pkg/services/dmsgsrv so the multi-service supervisor can host a
dmsg-server alongside dmsg-discovery and the rest of the deployment
services in one process. The standalone `skywire dmsg server start`
cobra command keeps working unchanged from an operator's POV — it
parses flags + config file into a dmsgsrv.Config and hands off to
Run().

Why dmsgsrv (not dmsgserver)
-----------------------------
The lower-level pkg/dmsg/dmsgserver already owns the Config struct,
the ServerAPI, NewServer, NormalizedDeployments, etc. The new
pkg/services/dmsgsrv is purely the run-loop wrapper that turns those
primitives into a pkg/services.Service. Two-package split keeps the
service abstraction (init() registration, Factory) out of the
existing low-level package's import surface.

Config
------
dmsgsrv.Config embeds dmsgserver.Config and adds the runtime knobs
that previously came from cobra flags only:

  type Config struct {
      dmsgserver.Config

      AuthPassphrase string `json:"auth_passphrase,omitempty"`
      PProfMode      string `json:"pprof_mode,omitempty"`
      PProfAddr      string `json:"pprof_addr,omitempty"`
      MetricsAddr    string `json:"metrics_addr,omitempty"`
  }

services.json block:

  {
    "type": "dmsg-server",
    "public_key": "...",
    "secret_key": "...",
    "discovery": "http://dmsg-discovery:9090",
    "discovery_dmsg": "dmsg://...",
    "local_address": ":8080",
    "public_address": "...:8080",
    "max_sessions": 2048,
    "enable_route_setup": true
  }

Behavioral parity
-----------------
- Same flag surface on `skywire dmsg server start`.
- Same NormalizedDeployments resolution: modern Dmsg.Deployments
  list wins, legacy Discovery / DiscoveryDmsg / PublicAddress are
  folded as fallback.
- Same dmsgfirst registration upgrade (when DiscoveryDmsg is set).
- Same /health endpoint over plain HTTP and DMSG (port 80).
- Same /debug/pprof over DMSG (port 81, gated by SurveyWhitelist).
- Same integrated route setup-node on dmsg port 36
  (cfg.EnableRouteSetup == true).
- Same dmsg-server transit set construction:
  union of (per-discovery servers, embedded keyring), minus self.

Registration
------------
cmd/svc/skywire-services/commands/root.go gets a side-effect import
of pkg/services/dmsgsrv, so:

  $ skywire svc run --list
  registered service types:
    dmsg-discovery
    dmsg-server
    noop

What's next
-----------
cmd/svc/ batch migration: tpd, sd, ar, rf, sn, tps, stun-server.
Then docker/docker-compose.yml collapses to one container running
`skywire svc run --config services.json` for CI e2e.
