+++ cmd/svc/skywire-services/commands/root.go
@@ -29,6 +29,7 @@ import (
 	// Side-effect imports register service factories with
 	// pkg/services so `skywire svc run` can dispatch them.
 	_ "github.com/skycoin/skywire/pkg/services/dmsgdisc"
+	_ "github.com/skycoin/skywire/pkg/services/dmsgsrv"
 	_ "github.com/skycoin/skywire/pkg/services/noop"
 )
